Crate Refactor
Refactor the crates into three separate crates:

A) serde_json_path
B) serde_json_path_macros
C) serde_json_path_core

The (C) core crate is needed because (A) and (B) will share dependencies
and thus would cause a circular dep in absence of (C). In general, all
of the core types and traits are moved from (A) to (C), while (A) still
holds some of the high-level user-facing types, as well as the parser.

(B) will still just be dedicated for the proc macro, but by moving the
core types into (C), (B) can now, hopefully, more reliably generate the
code needed for functions to compile correctly.

The code is compiling after this commit, however the functions have been
broken, and therefore there are a few tests failing.
