+++ serde_json_path/Cargo.toml
@@ -21,6 +21,7 @@ once_cell = { version = "1.17.1" }
 regex = "1.7.1"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+serde_json_path_core = { path = "../serde_json_path_core", version = "0.1.0" }
 serde_json_path_macros = { path = "../serde_json_path_macros", version = "0.1.0" }
 thiserror = "1.0"
 tracing = { version = "0.1", optional = true }
