+++ serde_json_path/src/parser/mod.rs
@@ -2,78 +2,16 @@ use nom::character::complete::char;
 use nom::combinator::all_consuming;
 use nom::error::VerboseError;
 use nom::{branch::alt, combinator::map, multi::many0, sequence::preceded, IResult};
-use serde_json::Value;
+use serde_json_path_core::spec::query::{PathKind, Query};
+use serde_json_path_core::spec::segment::PathSegment;
 
-use self::segment::{parse_segment, PathSegment};
+use self::segment::parse_segment;
 
 pub mod primitive;
 pub mod segment;
 pub mod selector;
 
 type PResult<'a, O> = IResult<&'a str, O, VerboseError<&'a str>>;
-
-pub trait Queryable {
-    fn query<'b>(&self, current: &'b Value, root: &'b Value) -> Vec<&'b Value>;
-}
-
-/// Represents a JSONPath expression
-#[derive(Debug, PartialEq, Clone, Default)]
-pub struct Query {
-    kind: PathKind,
-    pub segments: Vec<PathSegment>,
-}
-
-impl Query {
-    fn is_singular(&self) -> bool {
-        for s in &self.segments {
-            if s.is_descendent() {
-                return false;
-            }
-            if !s.segment.is_singular() {
-                return false;
-            }
-        }
-        return true;
-    }
-}
-
-impl std::fmt::Display for Query {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self.kind {
-            PathKind::Root => write!(f, "$")?,
-            PathKind::Current => write!(f, "@")?,
-        }
-        for s in &self.segments {
-            write!(f, "{s}")?;
-        }
-        Ok(())
-    }
-}
-
-#[derive(Debug, PartialEq, Clone, Default)]
-pub enum PathKind {
-    #[default]
-    Root,
-    Current,
-}
-
-impl Queryable for Query {
-    #[cfg_attr(feature = "trace", tracing::instrument(name = "Main Query", level = "trace", parent = None, ret))]
-    fn query<'b>(&self, current: &'b Value, root: &'b Value) -> Vec<&'b Value> {
-        let mut query = match self.kind {
-            PathKind::Root => vec![root],
-            PathKind::Current => vec![current],
-        };
-        for segment in &self.segments {
-            let mut new_query = Vec::new();
-            for q in &query {
-                new_query.append(&mut segment.query(q, root));
-            }
-            query = new_query;
-        }
-        query
-    }
-}
 
 #[cfg_attr(feature = "trace", tracing::instrument(level = "trace", parent = None, ret, err))]
 fn parse_path_segments(input: &str) -> PResult<Vec<PathSegment>> {
@@ -108,10 +46,10 @@ pub fn parse_path_main(input: &str) -> PResult<Query> {
 
 #[cfg(test)]
 mod tests {
-    use crate::parser::{
+    use serde_json_path_core::spec::{
+        query::PathKind,
         segment::Segment,
-        selector::{Name, Selector},
-        PathKind,
+        selector::{name::Name, Selector},
     };
 
     use super::{parse_path, parse_path_main};
