+++ serde_json_path_core/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "serde_json_path_core"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+
+[dependencies]
+inventory = { version = "0.3.4" }
+once_cell = "1"
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+thiserror = "1.0"
