+++ serde_json_path_core/src/spec/mod.rs
@@ -1,0 +1,4 @@
+pub mod functions;
+pub mod query;
+pub mod segment;
+pub mod selector;
