+++ serde_json_path_core/src/spec/selector/name.rs
@@ -1,0 +1,35 @@
+use serde_json::Value;
+
+use crate::spec::query::Queryable;
+
+#[derive(Debug, PartialEq, Clone)]
+pub struct Name(pub String);
+
+impl Name {
+    pub fn as_str(&self) -> &str {
+        &self.0
+    }
+}
+
+impl std::fmt::Display for Name {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "'{name}'", name = self.0)
+    }
+}
+
+impl Queryable for Name {
+    #[cfg_attr(feature = "trace", tracing::instrument(name = "Query Name", level = "trace", parent = None, ret))]
+    fn query<'b>(&self, current: &'b Value, _root: &'b Value) -> Vec<&'b Value> {
+        if let Some(obj) = current.as_object() {
+            obj.get(&self.0).into_iter().collect()
+        } else {
+            vec![]
+        }
+    }
+}
+
+impl From<&str> for Name {
+    fn from(s: &str) -> Self {
+        Self(s.to_owned())
+    }
+}
